requirements.txt
pandas
numpy
scikit-learn
joblib
spacy==3.8.13
streamlit==1.40.0
altair==4.2.2
plotly
python-dotenv
openai
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl